Add slice_assign autodiff tests and support for steps
Moved slice_assign autodiff tests to a new file and expanded them to cover step and negative step cases. Removed panic for slice steps in float_slice, enabling support for slice_assign with steps. Registered the new test suite in the test modules for both burn-autodiff and burn-candle.
